Move `execute` folder
