+++ client/web/src/enterprise/batches/batch-spec/execute/WorkspaceDetails.module.scss
