+++ client/web/src/enterprise/batches/batch-spec/execute/workspaces/Workspaces.module.scss
