+++ client/web/src/enterprise/batches/batch-spec/execute/workspaces/WorkspacesListItem.module.scss
