+++ client/web/src/enterprise/batches/batch-spec/execute/workspaces/WorkspacesListItem.story.tsx
@@ -2,8 +2,8 @@ import { useState } from 'react'
 
 import { storiesOf } from '@storybook/react'
 
-import { WebStory } from '../../../../components/WebStory'
-import { BatchSpecWorkspaceState } from '../../../../graphql-operations'
+import { WebStory } from '../../../../../components/WebStory'
+import { BatchSpecWorkspaceState } from '../../../../../graphql-operations'
 
 import { mockWorkspace } from './WorkspacesList.mock'
 import { WorkspacesListItem } from './WorkspacesListItem'
